EXPERIMENT Separate REST API+impl for proper code-coverage
