+++ servers/services/src/main/java/org/projectnessie/services/rest/ContentsResource.java
@@ -13,99 +13,62 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package org.projectnessie.services.rest;
 
 import java.security.Principal;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
 import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
 
-import org.projectnessie.api.ContentsApi;
+import org.projectnessie.api.ContentsRestApi;
 import org.projectnessie.error.NessieConflictException;
 import org.projectnessie.error.NessieNotFoundException;
 import org.projectnessie.model.CommitMeta;
 import org.projectnessie.model.Contents;
 import org.projectnessie.model.ContentsKey;
-import org.projectnessie.model.ImmutableMultiGetContentsResponse;
 import org.projectnessie.model.MultiGetContentsRequest;
 import org.projectnessie.model.MultiGetContentsResponse;
-import org.projectnessie.model.MultiGetContentsResponse.ContentsWithKey;
 import org.projectnessie.services.config.ServerConfig;
-import org.projectnessie.versioned.Delete;
-import org.projectnessie.versioned.Hash;
-import org.projectnessie.versioned.Key;
-import org.projectnessie.versioned.Put;
-import org.projectnessie.versioned.ReferenceNotFoundException;
+import org.projectnessie.services.impl.ContentsApiImpl;
 import org.projectnessie.versioned.VersionStore;
 
 /**
  * REST endpoint for contents.
  */
 @RequestScoped
-public class ContentsResource extends BaseResource implements ContentsApi {
+public class ContentsResource implements ContentsRestApi {
+
+  /**
+   * Delegate to (do not extend) the implementation for better code-coverage.
+   */
+  private final ContentsApiImpl delegate;
 
   @Inject
   public ContentsResource(ServerConfig config, Principal principal,
       VersionStore<Contents, CommitMeta> store) {
-    super(config, principal, store);
+    delegate = new ContentsApiImpl(config, principal, store);
   }
 
   @Override
   public Contents getContents(ContentsKey key, String incomingRef) throws NessieNotFoundException {
-    Hash ref = getHashOrThrow(incomingRef);
-    try {
-      Contents obj = getStore().getValue(ref, toKey(key));
-      if (obj != null) {
-        return obj;
-      }
-      throw new NessieNotFoundException("Requested contents do not exist for specified reference.");
-    } catch (ReferenceNotFoundException e) {
-      throw new NessieNotFoundException(String.format("Provided reference [%s] does not exist.", incomingRef), e);
-    }
-  }
-
-
-  @Override
-  public MultiGetContentsResponse getMultipleContents(String refName, MultiGetContentsRequest request)
-      throws NessieNotFoundException {
-    try {
-      Hash ref = getHashOrThrow(refName);
-      List<ContentsKey> externalKeys = request.getRequestedKeys();
-      List<Key> internalKeys = externalKeys.stream().map(ContentsResource::toKey).collect(Collectors.toList());
-      List<Optional<Contents>> values = getStore().getValues(ref, internalKeys);
-      List<ContentsWithKey> output = new ArrayList<>();
-
-      for (int i = 0; i < externalKeys.size(); i++) {
-        final int pos = i;
-        values.get(i).ifPresent(v -> output.add(ContentsWithKey.of(externalKeys.get(pos), v)));
-      }
-
-      return ImmutableMultiGetContentsResponse.builder().contents(output).build();
-    } catch (ReferenceNotFoundException ex) {
-      throw new NessieNotFoundException("Unable to find the requested ref.", ex);
-    }
+    return delegate.getContents(key, incomingRef);
   }
 
   @Override
-  public void setContents(ContentsKey key, String branch, String hash, String message, Contents contents)
-      throws NessieNotFoundException, NessieConflictException {
-    doOps(branch, hash, message, Arrays.asList(Put.of(toKey(key), contents)));
+  public MultiGetContentsResponse getMultipleContents(String refName,
+      MultiGetContentsRequest request) throws NessieNotFoundException {
+    return delegate.getMultipleContents(refName, request);
+  }
+
+  @Override
+  public void setContents(ContentsKey key, String branch, String hash, String message,
+      Contents contents) throws NessieNotFoundException, NessieConflictException {
+    delegate.setContents(key, branch, hash, message, contents);
   }
 
   @Override
   public void deleteContents(ContentsKey key, String branch, String hash, String message)
       throws NessieNotFoundException, NessieConflictException {
-    doOps(branch, hash, message, Arrays.asList(Delete.of(toKey(key))));
+    delegate.deleteContents(key, branch, hash, message);
   }
-
-  static Key toKey(ContentsKey key) {
-    return Key.of(key.getElements().toArray(new String[key.getElements().size()]));
-  }
-
 }
