+++ servers/services/src/main/java/org/projectnessie/services/rest/TreeResource.java
@@ -13,299 +13,120 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package org.projectnessie.services.rest;
 
 import java.security.Principal;
 import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
 
-import org.projectnessie.api.TreeApi;
+import org.projectnessie.api.TreeRestApi;
 import org.projectnessie.error.NessieConflictException;
 import org.projectnessie.error.NessieNotFoundException;
 import org.projectnessie.model.Branch;
 import org.projectnessie.model.CommitMeta;
 import org.projectnessie.model.Contents;
-import org.projectnessie.model.Contents.Type;
-import org.projectnessie.model.ContentsKey;
 import org.projectnessie.model.EntriesResponse;
-import org.projectnessie.model.ImmutableBranch;
-import org.projectnessie.model.ImmutableCommitMeta;
-import org.projectnessie.model.ImmutableHash;
-import org.projectnessie.model.ImmutableLogResponse;
-import org.projectnessie.model.ImmutableTag;
 import org.projectnessie.model.LogResponse;
 import org.projectnessie.model.Merge;
-import org.projectnessie.model.Operation;
 import org.projectnessie.model.Operations;
 import org.projectnessie.model.Reference;
 import org.projectnessie.model.Tag;
 import org.projectnessie.model.Transplant;
 import org.projectnessie.services.config.ServerConfig;
-import org.projectnessie.versioned.BranchName;
-import org.projectnessie.versioned.Delete;
-import org.projectnessie.versioned.Hash;
-import org.projectnessie.versioned.Key;
-import org.projectnessie.versioned.NamedRef;
-import org.projectnessie.versioned.Put;
-import org.projectnessie.versioned.Ref;
-import org.projectnessie.versioned.ReferenceAlreadyExistsException;
-import org.projectnessie.versioned.ReferenceConflictException;
-import org.projectnessie.versioned.ReferenceNotFoundException;
-import org.projectnessie.versioned.TagName;
-import org.projectnessie.versioned.Unchanged;
+import org.projectnessie.services.impl.TreeApiImpl;
 import org.projectnessie.versioned.VersionStore;
-import org.projectnessie.versioned.WithHash;
-
-import com.google.common.collect.ImmutableList;
 
 /**
  * REST endpoint for trees.
  */
 @RequestScoped
-public class TreeResource extends BaseResource implements TreeApi {
+public class TreeResource implements TreeRestApi {
 
-  private static final int MAX_COMMIT_LOG_ENTRIES = 250;
+  /**
+   * Delegate to (do not extend) the implementation for better code-coverage.
+   */
+  private final TreeApiImpl delegate;
 
   @Inject
   public TreeResource(ServerConfig config, Principal principal,
       VersionStore<Contents, CommitMeta> store) {
-    super(config, principal, store);
+    delegate = new TreeApiImpl(config, principal, store);
   }
 
   @Override
   public List<Reference> getAllReferences() {
-    try (Stream<WithHash<NamedRef>> str = getStore().getNamedRefs()) {
-      return str.map(TreeResource::makeNamedRef).collect(Collectors.toList());
-    }
+    return delegate.getAllReferences();
   }
 
   @Override
   public Reference getReferenceByName(String refName) throws NessieNotFoundException {
-    try {
-      return makeRef(getStore().toRef(refName));
-    } catch (ReferenceNotFoundException e) {
-      throw new NessieNotFoundException(String.format("Unable to find reference [%s].", refName), e);
-    }
+    return delegate.getReferenceByName(refName);
   }
 
   @Override
   public void createReference(Reference reference)
       throws NessieNotFoundException, NessieConflictException {
-    final NamedRef namedReference;
-    if (reference instanceof Branch) {
-      namedReference = BranchName.of(reference.getName());
-    } else if (reference instanceof Tag) {
-      namedReference = TagName.of(reference.getName());
-    } else {
-      throw new IllegalArgumentException("Only tag and branch references can be created");
-    }
-    createReference(namedReference, reference.getHash());
-  }
-
-  private void createReference(NamedRef reference, String hash) throws NessieNotFoundException, NessieConflictException {
-    try {
-      getStore().create(reference, toHash(hash, false));
-    } catch (ReferenceNotFoundException e) {
-      throw new NessieNotFoundException("Failure while searching for provided targeted hash.", e);
-    } catch (ReferenceAlreadyExistsException e) {
-      throw new NessieConflictException(String.format("A reference of name [%s] already exists.", reference.getName()), e);
-    }
+    delegate.createReference(reference);
   }
 
   @Override
   public Branch getDefaultBranch() throws NessieNotFoundException {
-    Reference r = getReferenceByName(getConfig().getDefaultBranch());
-    if (!(r instanceof Branch)) {
-      throw new IllegalStateException("Default branch isn't a branch");
-    }
-    return (Branch) r;
+    return delegate.getDefaultBranch();
   }
 
   @Override
   public void assignTag(String tagName, String expectedHash, Tag tag)
       throws NessieNotFoundException, NessieConflictException {
-    assignReference(TagName.of(tagName), expectedHash, tag.getHash());
+    delegate.assignTag(tagName, expectedHash, tag);
   }
 
   @Override
-  public void deleteTag(String tagName, String hash) throws NessieConflictException, NessieNotFoundException {
-    deleteReference(TagName.of(tagName), hash);
+  public void deleteTag(String tagName, String hash)
+      throws NessieConflictException, NessieNotFoundException {
+    delegate.deleteTag(tagName, hash);
   }
 
   @Override
   public void assignBranch(String branchName, String expectedHash, Branch branch)
       throws NessieNotFoundException, NessieConflictException {
-    assignReference(BranchName.of(branchName), expectedHash, branch.getHash());
+    delegate.assignBranch(branchName, expectedHash, branch);
   }
 
   @Override
-  public void deleteBranch(String branchName, String hash) throws NessieConflictException, NessieNotFoundException {
-    deleteReference(BranchName.of(branchName), hash);
+  public void deleteBranch(String branchName, String hash)
+      throws NessieConflictException, NessieNotFoundException {
+    delegate.deleteBranch(branchName, hash);
   }
 
   @Override
-  public LogResponse getCommitLog(String ref, Integer maxEntriesHint, String pageToken) throws NessieNotFoundException {
-    int max = Math.min(maxEntriesHint != null ? maxEntriesHint : MAX_COMMIT_LOG_ENTRIES, MAX_COMMIT_LOG_ENTRIES);
-    Hash startRef = getHashOrThrow(pageToken != null ? pageToken : ref);
-    try (Stream<ImmutableCommitMeta> s = getStore().getCommits(startRef).limit(max + 1)
-        .map(cwh -> cwh.getValue().toBuilder().hash(cwh.getHash().asString()).build())) {
-      List<CommitMeta> items = s.collect(Collectors.toList());
-      if (items.size() == max + 1) {
-        return ImmutableLogResponse.builder()
-            .addAllOperations(items.subList(0, max))
-            .hasMore(true).token(items.get(max).getHash())
-            .build();
-      }
-      return ImmutableLogResponse.builder()
-          .addAllOperations(items)
-          .build();
-    } catch (ReferenceNotFoundException e) {
-      throw new NessieNotFoundException(String.format("Unable to find the requested ref [%s].", ref), e);
-    }
+  public LogResponse getCommitLog(String ref, Integer maxEntriesHint, String pageToken)
+      throws NessieNotFoundException {
+    return delegate.getCommitLog(ref, maxEntriesHint, pageToken);
   }
 
   @Override
-  public void transplantCommitsIntoBranch(String branchName, String hash, String message, Transplant transplant)
+  public void transplantCommitsIntoBranch(String branchName, String hash, String message,
+      Transplant transplant) throws NessieNotFoundException, NessieConflictException {
+    delegate.transplantCommitsIntoBranch(branchName, hash, message, transplant);
+  }
+
+  @Override
+  public void mergeRefIntoBranch(String branchName, String hash, Merge merge)
       throws NessieNotFoundException, NessieConflictException {
-    try {
-      List<Hash> transplants;
-      try (Stream<Hash> s = transplant.getHashesToTransplant().stream().map(Hash::of)) {
-        transplants = s.collect(Collectors.toList());
-      }
-      getStore().transplant(BranchName.of(branchName), toHash(hash, true), transplants);
-    } catch (ReferenceNotFoundException e) {
-      throw new NessieNotFoundException(
-          String.format("Unable to find the requested branch we're transplanting to of [%s].", branchName), e);
-    } catch (ReferenceConflictException e) {
-      throw new NessieConflictException(
-          String.format("The hash provided %s does not match the current status of the branch %s.",
-              hash, branchName), e);
-    }
+    delegate.mergeRefIntoBranch(branchName, hash, merge);
   }
 
   @Override
-  public void mergeRefIntoBranch(String branchName, String hash, Merge merge) throws NessieNotFoundException, NessieConflictException {
-    try {
-      getStore().merge(toHash(merge.getFromHash(), true).get(), BranchName.of(branchName), toHash(hash, true));
-    } catch (ReferenceNotFoundException e) {
-      throw new NessieNotFoundException(String.format("At least one of the references provided does not exist."), e);
-    } catch (ReferenceConflictException e) {
-      throw new NessieConflictException(
-          String.format("The branch [%s] does not have the expected hash [%s].", branchName, hash), e);
-    }
+  public EntriesResponse getEntries(String refName, Integer maxEntriesHint, String pageToken)
+      throws NessieNotFoundException {
+    return delegate.getEntries(refName, maxEntriesHint, pageToken);
   }
 
   @Override
-  public EntriesResponse getEntries(String refName, Integer maxEntriesHint, String pageToken) throws NessieNotFoundException {
-    final Hash hash = getHashOrThrow(refName);
-    // TODO Implement paging. At the moment, we do not expect that many keys/entries to be returned.
-    //  So the size of the whole result is probably reasonable and unlikely to "kill" either the
-    //  server or client. We have to figure out _how_ to implement paging for keys/entries, i.e.
-    //  whether we shall just do the whole computation for a specific hash for every page or have
-    //  a more sophisticated approach, potentially with support from the (tiered-)version-store.
-    try {
-      List<EntriesResponse.Entry> entries;
-      try (Stream<EntriesResponse.Entry> s = getStore().getKeys(hash)
-          .map(key -> EntriesResponse.Entry.builder().name(fromKey(key)).type(Type.UNKNOWN).build())) {
-        entries = s.collect(ImmutableList.toImmutableList());
-      }
-      return EntriesResponse.builder().addAllEntries(entries).build();
-    } catch (ReferenceNotFoundException e) {
-      throw new NessieNotFoundException(String.format("Unable to find the reference [%s].", refName), e);
-    }
-  }
-
-  @Override
-  public void commitMultipleOperations(String branch, String hash, String message, Operations operations)
-      throws NessieNotFoundException, NessieConflictException {
-    List<org.projectnessie.versioned.Operation<Contents>> ops = operations.getOperations()
-        .stream()
-        .map(TreeResource::toOp)
-        .collect(ImmutableList.toImmutableList());
-    doOps(branch, hash, message, ops);
-  }
-
-  private static Optional<Hash> toHash(String hash, boolean required) throws NessieConflictException {
-    if (hash == null || hash.isEmpty()) {
-      if (required) {
-        throw new NessieConflictException("Must provide expected hash value for operation.");
-      }
-      return Optional.empty();
-    }
-    return Optional.of(Hash.of(hash));
-  }
-
-  private void deleteReference(NamedRef name, String hash) throws NessieConflictException, NessieNotFoundException {
-    try {
-      getStore().delete(name, toHash(hash, true));
-    } catch (ReferenceNotFoundException e) {
-      throw new NessieNotFoundException(String.format("Unable to find reference [%s] to delete.", name.getName()), e);
-    } catch (ReferenceConflictException e) {
-      throw new NessieConflictException(
-          String.format("The hash provided %s does not match the current status of the reference %s.",
-              hash, name.getName()), e);
-    }
-  }
-
-  private void assignReference(NamedRef ref, String oldHash, String newHash)
-      throws NessieNotFoundException, NessieConflictException {
-    try {
-      WithHash<Ref> resolved = getStore().toRef(ref.getName());
-      Ref resolvedRef = resolved.getValue();
-      if (resolvedRef instanceof NamedRef) {
-        getStore().assign((NamedRef) resolvedRef, toHash(oldHash, true), toHash(newHash, true)
-            .orElseThrow(() -> new NessieConflictException("Must provide target hash value for operation.")));
-      } else {
-        throw new IllegalArgumentException("Can only assign branch and tag types.");
-      }
-    } catch (ReferenceNotFoundException e) {
-      throw new NessieNotFoundException("Unable to find a ref or hash provided.", e);
-    } catch (ReferenceConflictException e) {
-      throw new NessieConflictException(
-          String.format("The hash provided %s does not match the current status of the reference %s.",
-              oldHash, ref), e);
-    }
-  }
-
-  private static ContentsKey fromKey(Key key) {
-    return ContentsKey.of(key.getElements());
-  }
-
-  private static Reference makeNamedRef(WithHash<NamedRef> refWithHash) {
-    return makeRef(refWithHash);
-  }
-
-  private static Reference makeRef(WithHash<? extends Ref> refWithHash) {
-    Ref ref = refWithHash.getValue();
-    //todo do we want to send back Hash object or the string. I don't want internal API escaping so maybe an external representation of hash
-    if (ref instanceof TagName) {
-      return ImmutableTag.builder().name(((NamedRef)ref).getName()).hash(refWithHash.getHash().asString()).build();
-    } else if (ref instanceof BranchName) {
-      return ImmutableBranch.builder().name(((NamedRef)ref).getName()).hash(refWithHash.getHash().asString()).build();
-    } else if (ref instanceof Hash) {
-      String hash = refWithHash.getHash().asString();
-      return ImmutableHash.builder().name(hash).build();
-    } else {
-      throw new UnsupportedOperationException("only converting tags or branches"); //todo
-    }
-  }
-
-  private static org.projectnessie.versioned.Operation<Contents> toOp(Operation o) {
-    Key key = Key.of(o.getKey().getElements().toArray(new String[0]));
-    if (o instanceof Operation.Delete) {
-      return Delete.of(key);
-    } else if (o instanceof Operation.Put) {
-      return Put.of(key, ((Operation.Put)o).getContents());
-    } else if (o instanceof Operation.Unchanged) {
-      return Unchanged.of(key);
-    } else {
-      throw new IllegalStateException("Unknown operation " + o);
-    }
+  public void commitMultipleOperations(String branch, String hash, String message,
+      Operations operations) throws NessieNotFoundException, NessieConflictException {
+    delegate.commitMultipleOperations(branch, hash, message, operations);
   }
 }
